requirements.txt
flask
torch==2.3.1+cpu
sentence-transformers
faiss-cpu
numpy